requirements.txt
pymupdf==1.25.3
chromadb==0.6.3
google-generativeai==0.8.4
python-dotenv==1.1.0
rich==14.0.0
